+++ app/src/docs/_props/loveship/components/inputs/numericInput.props.ts
@@ -27,7 +27,6 @@ const NumericInputDoc = new DocBuilder<NumericInputProps & NumericInputMods>({ n
     })
     .prop('min', { examples: [0, 10], defaultValue: 0 })
     .prop('max', { examples: [20, 50] })
-    .prop('mode', { examples: ['form', 'cell'] })
     .prop('align', { examples: ['left', 'right'] })
     .prop('disableArrows', { examples: [true, false] })
     .prop('disableLocaleFormatting', { defaultValue: false, examples: [true, false] })